Move plGBufferGroup and plVertCoder to plDrawable.
