+++ Sources/Plasma/PubUtilLib/plDrawable/plAccessGeometry.cpp
@@ -57,7 +57,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plAccessSnapShot.h"
 
 // For dipping directly into device buffers.
-#include "plPipeline/plGBufferGroup.h"
+#include "plGBufferGroup.h"
 #include "plPipeline/hsGDeviceRef.h"
 #include "plPipeline.h"
