+++ Sources/Plasma/PubUtilLib/plDrawable/plGBufferGroup.cpp
@@ -56,9 +56,9 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "hsStream.h"
 
 #include "plSurface/hsGMaterial.h"
-#include "plDrawable/plGeometrySpan.h"
+#include "plGeometrySpan.h"
 #include "plPipeline.h"
-#include "hsGDeviceRef.h"
+#include "plPipeline/hsGDeviceRef.h"
 #include "plProfile.h"
 
 #include "plVertCoder.h"
