+++ Sources/Plasma/PubUtilLib/plPipeline/CMakeLists.txt
@@ -18,14 +18,12 @@ set(plPipeline_SOURCES
     plDTProgressMgr.cpp
     plDynamicEnvMap.cpp
     plFogEnvironment.cpp
-    plGBufferGroup.cpp
     plPlates.cpp
     plRenderTarget.cpp
     plStatusLogDrawer.cpp
     plTextFont.cpp
     plTextGenerator.cpp
     plTransitionMgr.cpp
-    plVertCoder.cpp
 )
 
 set(plDXPipeline_SOURCES
@@ -55,7 +53,6 @@ set(plPipeline_HEADERS
     plDTProgressMgr.h
     plDynamicEnvMap.h
     plFogEnvironment.h
-    plGBufferGroup.h
     plPipeDebugFlags.h
     plPipelineCreatable.h
     plPipelineCreate.h
@@ -66,7 +63,6 @@ set(plPipeline_HEADERS
     plTextFont.h
     plTextGenerator.h
     plTransitionMgr.h
-    plVertCoder.h
 )
 
 set(plDXPipeline_HEADERS
