+++ Sources/Plasma/PubUtilLib/plPipeline/DX/plDXVertexShader.cpp
@@ -49,7 +49,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 #include "plSurface/plShader.h"
 
-#include "plPipeline/plGBufferGroup.h"
+#include "plDrawable/plGBufferGroup.h"
 #include "plDXPipeline.h"
 
 plDXVertexShader::plDXVertexShader(plShader* owner)
